Make detection of amd64 vs. arm64 in format script automatic on Linux (#161)
* Make detection of amd64 vs. arm64 automatic on Linux

* Only install buildifier if it has not been installed already
In case someone wants to run format.sh multiple times on the same
system.

* Update buildifier from version 7.3.1 to 8.0.3
Running the newer version causes a few lines to be added to
5 .bazel files.

+++ format.sh
@@ -14,7 +14,16 @@
 # and similarly for other platforms.
 
 # Possible values: {linux, darwin}_{amd64, arm64}.
-PLATFORM=${PLATFORM:-linux-amd64}
+which dpkg
+exit_status=$?
+if [ ${exit_status} -eq 0 ]
+then
+    ARCH=`dpkg --print-architecture`
+    PLATFORM=${PLATFORM:-linux-${ARCH}}
+else
+    # Guess amd64
+    PLATFORM=${PLATFORM:-linux-amd64}
+fi
 
 # Only files with these extensions will be formatted by clang-format.
 CLANG_FORMAT_EXTENSIONS="cc|h|proto"
@@ -25,8 +34,11 @@ find . -not -path "./third_party/**" \
   | xargs clang-format --verbose -style=google -i
 
 # Run buildifier (Bazel file formatter).
-BUILDIFIER="buildifier-$PLATFORM"
-wget "https://github.com/bazelbuild/buildtools/releases/download/v7.3.1/$BUILDIFIER"
-mv $BUILDIFIER buildifier
-chmod +x buildifier
+if [ ! -e buildifier ]
+then
+    BUILDIFIER="buildifier-$PLATFORM"
+    wget "https://github.com/bazelbuild/buildtools/releases/download/v8.0.3/$BUILDIFIER"
+    mv $BUILDIFIER buildifier
+    chmod +x buildifier
+fi
 ./buildifier --lint=fix -r .
